Broke apart language definitions from the pain patterns file.
