+++ scripts/pattern-detection/utils.mjs
@@ -1,0 +1,15 @@
+/**
+ * Builds a regex alternation group (?:key1|key2|...) from map keys.
+ */
+export function buildAlternationGroup(map) {
+  if (!map || typeof map !== "object") return "";
+
+  // Escape basic regex chars
+  const escapedKeys = Object.keys(map)
+    .map((key) => key.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"))
+    .sort((a, b) => b.length - a.length);
+
+  if (escapedKeys.length === 0) return "";
+
+  return `(?:${escapedKeys.join("|")})`;
+}
